+++ soft/main/hal/marg.cpp
@@ -23,17 +23,17 @@ esp_err_t Marg::init(void)
     return ESP_FAIL;
 }
 
-esp_err_t Marg::read_acc(float * acc_x, float * acc_y, float * acc_z)
+esp_err_t Marg::read_acc(double * acc_x, double * acc_y, double * acc_z)
 {
     return _mpu->read_acc(acc_x, acc_y, acc_z);
 }
 
-esp_err_t Marg::read_gyro(float * gyro_x, float * gyro_y, float * gyro_z)
+esp_err_t Marg::read_gyro(double * gyro_x, double * gyro_y, double * gyro_z)
 {
     return _mpu->read_gyro(gyro_x, gyro_y, gyro_z);
 }
 
-esp_err_t Marg::read_mag(float * mag_x, float * mag_y, float * mag_z)
+esp_err_t Marg::read_mag(double * mag_x, double * mag_y, double * mag_z)
 {
     return _hmc->read_mag(mag_x, mag_y, mag_z);
 }
